Accomodate 'nasl' gem interface change (5.2 support)
With Nessus 5.2, function calls can now be any lvalue instead of only
an identifier. For example, this is now valid:
  foo['bar']['baz']();

This is also valid (and is syntactic sugar for the above, and is
therefore generally invisible to Pedant checks):
  foo.bar.baz();

The interface between the 'nasl' gem and Pedant has changed slightly
as a result: The "name" property of a Nasl::Call node now refers to a
Nasl::Lvalue instead of a Nasl::Identifier.