Revert "Add first iteration of ScratchpadQueriesUseConcatenation"
This reverts commit 02faadd43c8b4c0b0d76ec38546310b6844bd0fa.
This check is being excluded from the v0.0.5 release because it fires on a lot
of reasonable uses, such as:

  function insert_into_key_value(table, key, value)
  {
    query_scratchpad("INSERT INTO " + table + " VALUES (?, ?);", key, value);
  }

In this case, 'table' cannot be parameterized so it has to be concatted. It may
be worth generating a control-flow graph or some kind of taint tracking for
Pedant to do a better version of this check in the future.  If we later include
this check in its current form, it should be a 'warn' instead of a 'fail'.