Unbreak tests by lending dep. code to "test mode"
Tests broke as a result of the previous commit, sorry. In "file mode" (what
users use when actually checking files), we use a solid algorithm for running
checks as their dependencies get satisfied, stopping if we deadlock.

In "test mode", we were just running all checks that a tested check depends on,
whether those dependent checks should run or not.

This was never a problem before this change, from the previous commit:
  --- a/lib/pedant/checks/files_parse_without_errors.rb
  +++ b/lib/pedant/checks/files_parse_without_errors.rb
  @@ -30,7 +30,7 @@ module Pedant
         super + [:file_mode, :base, :main]
       end

  -    def provides
  +    def self.provides
         super + [:codes, :trees]
       end

In the ad-hoc dependency resolution, we ran all checks that a tested check
depends on: after the above, both FilesParseWithoutErrors and ParseTestCode
would run (we never checked if they *should*), and FilesParseWithoutErrors went
second and clobbered the results of ParseTestCode.
Because FilesParseWithoutErrors used to not be a candidate for providing :codes
and :trees, it didn't run during "test mode" and never clobbered anything.
However, because it had so few dependencies it would always run first in "file
mode", so everything just happened to work.